@@ -1,14 +1,13 @@
 package com.youarethomas.arborealis.block_entities;
 
 import com.youarethomas.arborealis.Arborealis;
-import net.fabricmc.fabric.api.block.entity.BlockEntityClientSerializable;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.util.math.BlockPos;
 
-public class HollowedLogEntity extends BlockEntity implements BlockEntityClientSerializable {
+public class HollowedLogEntity extends BlockEntity {
 
     private String itemID = "arborealis:item/tree_core";
 
@@ -29,12 +28,10 @@ public class HollowedLogEntity extends BlockEntity implements BlockEntityClientS
 
     // Serialize the BlockEntity - storing data
     @Override
-    public NbtCompound writeNbt(NbtCompound tag) {
+    public void writeNbt(NbtCompound tag) {
         super.writeNbt(tag);
 
         tag.putString("item_id", itemID);
-
-        return tag;
     }
 
     // Deserialize the BlockEntity - retrieving data
@@ -43,17 +40,6 @@ public class HollowedLogEntity extends BlockEntity implements BlockEntityClientS
         super.readNbt(tag);
 
         itemID = tag.getString("item_id");
-    }
-
-    @Override
-    public void fromClientTag(NbtCompound tag) {
-        readNbt(tag);
-        updateListeners();
-    }
-
-    @Override
-    public NbtCompound toClientTag(NbtCompound tag) {
-        return writeNbt(tag);
     }
 
     private void updateListeners() {
