@@ -14,12 +14,13 @@ import net.minecraft.client.render.RenderLayers;
 import net.minecraft.client.texture.Sprite;
 import net.minecraft.client.util.SpriteIdentifier;
 import net.minecraft.util.math.Direction;
+import net.minecraft.world.World;
 
 import java.util.HashMap;
 import java.util.function.Function;
 
+@Environment(EnvType.CLIENT)
 public class DynamicCuboid {
-
     private static final float PIXEL_SIZE = 0.0625f; // Which is 1/16
 
     public HashMap<Direction, SpriteIdentifier> spriteIds = new HashMap<>();
@@ -86,7 +87,6 @@ public class DynamicCuboid {
         spriteIds.replace(side, spriteIdentifier);
     }
 
-    @Environment(EnvType.CLIENT)
     public void create(QuadEmitter emitter, Function<SpriteIdentifier, Sprite> textureGetter) {
 
         /* So emitters are kinda complicated:
@@ -94,68 +94,66 @@ public class DynamicCuboid {
            the left and bottom numbers are smaller than the right and top numbers, otherwise you'll
            get inverted planes (texture on the wrong side). */
 
-        if (MinecraftClient.getInstance().world.isClient()) {
-            for (Direction direction : Direction.values()) {
-                // Overlay things
-                int overlayColour = -1;
-                if (overlays.containsKey(direction)) {
-                    overlayColour = overlays.get(direction);
-                }
-                // Emissive things
-                boolean isEmissive = false;
-                if (emissives.containsKey(direction)) {
-                    isEmissive = emissives.get(direction);
-                }
+        for (Direction direction : Direction.values()) {
+            // Overlay things
+            int overlayColour = -1;
+            if (overlays.containsKey(direction)) {
+                overlayColour = overlays.get(direction);
+            }
+            // Emissive things
+            boolean isEmissive = false;
+            if (emissives.containsKey(direction)) {
+                isEmissive = emissives.get(direction);
+            }
 
-                switch (direction) {
-                    case NORTH -> {
-                        emitter.square(direction, 1f - ((x + xSize) * PIXEL_SIZE), y * PIXEL_SIZE, 1f - (x * PIXEL_SIZE), (y + ySize) * PIXEL_SIZE, z * PIXEL_SIZE);
-                        emitter.spriteBake(0, textureGetter.apply(spriteIds.get(Direction.NORTH)), MutableQuadView.BAKE_LOCK_UV);
-                        emitter.spriteColor(0, overlayColour, overlayColour, overlayColour, overlayColour);
-                        if (isEmissive)
-                            emitter.material(RendererAccess.INSTANCE.getRenderer().materialFinder().emissive(0, true).find());
-                        emitter.emit();
-                    }
-                    case SOUTH -> {
-                        emitter.square(direction, x * PIXEL_SIZE, y * PIXEL_SIZE, (x + xSize) * PIXEL_SIZE, (y + ySize) * PIXEL_SIZE, 1f - ((z + zSize) * PIXEL_SIZE));
-                        emitter.spriteBake(0, textureGetter.apply(spriteIds.get(Direction.SOUTH)), MutableQuadView.BAKE_LOCK_UV);
-                        emitter.spriteColor(0, overlayColour, overlayColour, overlayColour, overlayColour);
-                        if (isEmissive)
-                            emitter.material(RendererAccess.INSTANCE.getRenderer().materialFinder().emissive(0, true).find());
-                        emitter.emit();
-                    }
-                    case EAST -> {
-                        emitter.square(direction, 1f - ((z + zSize) * PIXEL_SIZE), y * PIXEL_SIZE, 1f - (z * PIXEL_SIZE), (y + ySize) * PIXEL_SIZE, 1f - ((x + xSize) * PIXEL_SIZE));
-                        emitter.spriteBake(0, textureGetter.apply(spriteIds.get(Direction.EAST)), MutableQuadView.BAKE_LOCK_UV);
-                        emitter.spriteColor(0, overlayColour, overlayColour, overlayColour, overlayColour);
-                        if (isEmissive)
-                            emitter.material(RendererAccess.INSTANCE.getRenderer().materialFinder().emissive(0, true).find());
-                        emitter.emit();
-                    }
-                    case WEST -> {
-                        emitter.square(direction, z * PIXEL_SIZE, y * PIXEL_SIZE, (z + zSize) * PIXEL_SIZE, (y + ySize) * PIXEL_SIZE, x * PIXEL_SIZE);
-                        emitter.spriteBake(0, textureGetter.apply(spriteIds.get(Direction.WEST)), MutableQuadView.BAKE_LOCK_UV);
-                        emitter.spriteColor(0, overlayColour, overlayColour, overlayColour, overlayColour);
-                        if (isEmissive)
-                            emitter.material(RendererAccess.INSTANCE.getRenderer().materialFinder().emissive(0, true).find());
-                        emitter.emit();
-                    }
-                    case UP -> {
-                        emitter.square(direction, x * PIXEL_SIZE, 1f - ((z + zSize) * PIXEL_SIZE), (x + xSize) * PIXEL_SIZE, 1f - (z * PIXEL_SIZE), 1f - ((y + ySize) * PIXEL_SIZE));
-                        emitter.spriteBake(0, textureGetter.apply(spriteIds.get(Direction.UP)), MutableQuadView.BAKE_LOCK_UV);
-                        emitter.spriteColor(0, overlayColour, overlayColour, overlayColour, overlayColour);
-                        if (isEmissive)
-                            emitter.material(RendererAccess.INSTANCE.getRenderer().materialFinder().emissive(0, true).find());
-                        emitter.emit();
-                    }
-                    case DOWN -> {
-                        emitter.square(direction, x * PIXEL_SIZE, z * PIXEL_SIZE, (x + xSize) * PIXEL_SIZE, (z + zSize) * PIXEL_SIZE, y * PIXEL_SIZE);
-                        emitter.spriteBake(0, textureGetter.apply(spriteIds.get(Direction.DOWN)), MutableQuadView.BAKE_LOCK_UV);
-                        emitter.spriteColor(0, overlayColour, overlayColour, overlayColour, overlayColour);
-                        if (isEmissive)
-                            emitter.material(RendererAccess.INSTANCE.getRenderer().materialFinder().emissive(0, true).find());
-                        emitter.emit();
-                    }
+            switch (direction) {
+                case NORTH -> {
+                    emitter.square(direction, 1f - ((x + xSize) * PIXEL_SIZE), y * PIXEL_SIZE, 1f - (x * PIXEL_SIZE), (y + ySize) * PIXEL_SIZE, z * PIXEL_SIZE);
+                    emitter.spriteBake(0, textureGetter.apply(spriteIds.get(Direction.NORTH)), MutableQuadView.BAKE_LOCK_UV);
+                    emitter.spriteColor(0, overlayColour, overlayColour, overlayColour, overlayColour);
+                    if (isEmissive)
+                        emitter.material(RendererAccess.INSTANCE.getRenderer().materialFinder().emissive(0, true).find());
+                    emitter.emit();
+                }
+                case SOUTH -> {
+                    emitter.square(direction, x * PIXEL_SIZE, y * PIXEL_SIZE, (x + xSize) * PIXEL_SIZE, (y + ySize) * PIXEL_SIZE, 1f - ((z + zSize) * PIXEL_SIZE));
+                    emitter.spriteBake(0, textureGetter.apply(spriteIds.get(Direction.SOUTH)), MutableQuadView.BAKE_LOCK_UV);
+                    emitter.spriteColor(0, overlayColour, overlayColour, overlayColour, overlayColour);
+                    if (isEmissive)
+                        emitter.material(RendererAccess.INSTANCE.getRenderer().materialFinder().emissive(0, true).find());
+                    emitter.emit();
+                }
+                case EAST -> {
+                    emitter.square(direction, 1f - ((z + zSize) * PIXEL_SIZE), y * PIXEL_SIZE, 1f - (z * PIXEL_SIZE), (y + ySize) * PIXEL_SIZE, 1f - ((x + xSize) * PIXEL_SIZE));
+                    emitter.spriteBake(0, textureGetter.apply(spriteIds.get(Direction.EAST)), MutableQuadView.BAKE_LOCK_UV);
+                    emitter.spriteColor(0, overlayColour, overlayColour, overlayColour, overlayColour);
+                    if (isEmissive)
+                        emitter.material(RendererAccess.INSTANCE.getRenderer().materialFinder().emissive(0, true).find());
+                    emitter.emit();
+                }
+                case WEST -> {
+                    emitter.square(direction, z * PIXEL_SIZE, y * PIXEL_SIZE, (z + zSize) * PIXEL_SIZE, (y + ySize) * PIXEL_SIZE, x * PIXEL_SIZE);
+                    emitter.spriteBake(0, textureGetter.apply(spriteIds.get(Direction.WEST)), MutableQuadView.BAKE_LOCK_UV);
+                    emitter.spriteColor(0, overlayColour, overlayColour, overlayColour, overlayColour);
+                    if (isEmissive)
+                        emitter.material(RendererAccess.INSTANCE.getRenderer().materialFinder().emissive(0, true).find());
+                    emitter.emit();
+                }
+                case UP -> {
+                    emitter.square(direction, x * PIXEL_SIZE, 1f - ((z + zSize) * PIXEL_SIZE), (x + xSize) * PIXEL_SIZE, 1f - (z * PIXEL_SIZE), 1f - ((y + ySize) * PIXEL_SIZE));
+                    emitter.spriteBake(0, textureGetter.apply(spriteIds.get(Direction.UP)), MutableQuadView.BAKE_LOCK_UV);
+                    emitter.spriteColor(0, overlayColour, overlayColour, overlayColour, overlayColour);
+                    if (isEmissive)
+                        emitter.material(RendererAccess.INSTANCE.getRenderer().materialFinder().emissive(0, true).find());
+                    emitter.emit();
+                }
+                case DOWN -> {
+                    emitter.square(direction, x * PIXEL_SIZE, z * PIXEL_SIZE, (x + xSize) * PIXEL_SIZE, (z + zSize) * PIXEL_SIZE, y * PIXEL_SIZE);
+                    emitter.spriteBake(0, textureGetter.apply(spriteIds.get(Direction.DOWN)), MutableQuadView.BAKE_LOCK_UV);
+                    emitter.spriteColor(0, overlayColour, overlayColour, overlayColour, overlayColour);
+                    if (isEmissive)
+                        emitter.material(RendererAccess.INSTANCE.getRenderer().materialFinder().emissive(0, true).find());
+                    emitter.emit();
                 }
             }
         }
